refactor resource definitions
