+++ extensions/azurePublishNew/src/node/azureResources/azureFunction.ts
@@ -2,18 +2,9 @@
 // Licensed under the MIT License.
 
 import { parseRuntimeKey } from '../../../../../Composer/packages/lib/shared';
-import { AZURE_HOSTING_GROUP_NAME } from '../getResources';
-import { ProvisionMethod, ProvisionWorkingSet, ResourceDefinition, ResourceProvisionService } from '../types';
+import { ProvisionMethod, ProvisionWorkingSet, ResourceProvisionService } from '../types';
 
 import { AzureFunctionConfig } from './types';
-
-export const azureFunctionsDefinition: ResourceDefinition = {
-  key: 'azureFunctions',
-  description: 'Azure Functions hosting your bot services.',
-  text: 'Azure Functions',
-  tier: 'S1 Standard',
-  group: AZURE_HOSTING_GROUP_NAME,
-};
 
 const getAzureFunctionsProvisionMethod = (): ProvisionMethod => {
   return (config: AzureFunctionConfig, workingSet: ProvisionWorkingSet): Promise<ProvisionWorkingSet> => {
